CellQuadrature: add compilation unit
