+++ CMakeLists_files.cmake
@@ -45,6 +45,7 @@ list (APPEND MAIN_SOURCE_FILES
   opm/grid/common/WellConnections.cpp
   opm/grid/common/ZoltanGraphFunctions.cpp
   opm/grid/common/ZoltanPartition.cpp
+  opm/grid/CellQuadrature.cpp
   opm/grid/GridHelpers.cpp
   opm/grid/GridManager.cpp
   opm/grid/GridUtilities.cpp
